website/flask_server/requirements.txt
beautifulsoup4==4.12.3
blinker==1.7.0
bs4==0.0.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cryptography==3.4.8
distro==1.7.0
distro-info
Flask==3.0.0
Flask-Cors==4.0.0
gunicorn==21.2.0
httplib2==0.20.2
idna==3.6
importlib-metadata==4.6.4
itsdangerous==2.1.2
Jinja2==3.1.3
keyring==23.5.0
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
MarkupSafe==2.1.3
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
packaging==23.2
PyJWT==2.3.0
pyparsing==2.4.7
PyYAML
requests==2.31.0
SecretStorage==3.3.1
six==1.16.0
soupsieve==2.5
urllib3==2.1.0
Werkzeug==3.0.1
zipp==1.0.0

# Ajout de dotenv et pytest
python-dotenv==1.0.1
pytest==8.0.0
